Refactor some functionality into utils module for external use
